Rename folder `experimental` -> `stateless`
