+++ packages/itwin/tree-widget/src/components/trees/stateless/common/UseHierarchyVisibility.ts
