+++ packages/itwin/tree-widget/src/components/trees/stateless/common/UseMultiCheckboxHandler.ts
