+++ packages/itwin/tree-widget/src/components/trees/stateless/common/components/VisibilityTreeNodeCheckbox.tsx
