+++ packages/itwin/tree-widget/src/components/trees/stateless/common/components/VisibilityTreeRenderer.scss
